Move updateOrganizationContext mutation to organization.graphql
Keep console base.graphql with only an empty type Mutation,
consistent with connect and trust base files.
